user_management/requirements.txt
Django>=4.0,<5.0
djangorestframework>=3.14
djangorestframework-simplejwt>=5.0
psycopg2-binary
channels
channels-redis
daphne
redis
django-redis

